workflows/marketing-pro-workflow/.claude/skills/qr-code-generator/scripts/requirements.txt
qrcode>=7.3.1
pillow>=9.0.0
